[Points] Fix events.data_indices for ActionType.ADDED event when adding single point (#7983)
# Description

(Bugfix) Events for adding/removing points have a data_indices property.
This works as expected for ActionType.REMOVING, .REMOVED, and .ADDING.

For ActionType.ADDED, however, a bug occured when only a single point is
added (which is the standard case in the GUI). The intention in the
original code was to query the number of points via len(coords), but
this does not produce the desired result when the coords variable is
just a one-dimensional point.

The same np.atleast_2d function that I utilised for the fix was already
used in _set_data a few lines above; I hope this fix is trivial enough
to pass without test case.

## Edit: additional info

*Copied by @jni from @psobolewskiPhD's comments ([1][1], [2][2]) below.*

For anyone wanting a bit more details, for a single point, `coords` is
e.g. `[18, 18]` but for multiple, it's e.g. `[[18, 18], [18, 18]]`. So
for the single point case, the `len` returns the number of axes, rather
than the number of points, so the `data_indexes` ends up being a e.g.
(-2, -1), rather than (-1,). For multiple points it works fine of
course. So the fix here is absolutely correct.

I looked at our tests and lo and behold, we *only* test the multiple
point case, which is why this bug sailed through tests:

https://github.com/napari/napari/blob/18b263bb73eb2e87938a3f0e9af9e65d2d9ab6a5/src/napari/layers/points/_tests/test_points.py#L1189

[…]

I parameterized test_add_points_direct to also check adding a single
point.
This fails on main but is fixed by this PR.

[1]: https://github.com/napari/napari/pull/7983#issuecomment-2927268669
[2]: https://github.com/napari/napari/pull/7983#pullrequestreview-2885768453

---------

Co-authored-by: Peter Sobolewski <pete.sd@gmail.com>