Optional rotation handle for selection box overlay + simplify inheritance for Vispy overlays (#7958)
# References and relevant issues

This PR is a follow-up discussion with @brisvag in reference to a zulip
discussion about using overlays for ROI management in image layers
(which also triggered issue #7931).

# Description

It adds an optional boolean flag `rotation` in the `InteractionBox`
compound to allow the enabling/disabling of the rotation handle. By
default this is kept to `True` to maintain backward compatibility with
other existing overlays that take advantage of such compound. Some minor
changes are applied to other transform box relevant functions to ensure
correct behavior.

Additionally, @brisvag identified a problem in the Vispy overlays
inheritance to ensure that the correct overlay class is built when
updating the `overlay_to_visual` dictionary.

An example is provided in `examples\dev\interaction_box.py`.

The usage of the example is dual fold: it provides a way to demonstrate
how to inherit from the selection box overlay to create a custom one,
and it uses such custom overlay in conjunction with two mouse event
handles to drag and drop the overlay handles on top of the layer to
simulate the selection of a sub-region of an image layer.


https://github.com/user-attachments/assets/3724fd42-523b-4fae-82c6-6e09637f8fb2

Furthermore, I took the liberty of making some better type annotations
to `InteractionBox`.

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Tim Monko <timmonko@gmail.com>