Fix moving of first/last vertex of polygons added in ring mode (#7942)
# References and relevant issues

Address problem pointed in
https://github.com/napari/napari/issues/5770#issuecomment-2887792751

# Description

If a polygon is added with same point as start and begin (implicit
close) ring (same as geoJSON) and we use direct mode for edit, we move
only one of this point, effectively adding an additional point to the
polygon that is unexpected.

This PR fixes this, by moving both points at the same time.