Remove the reset_scroll_progress action and keybinding (#7990)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/7232

# Description
There is a strange `Reset scroll` viewer keybinding, which isn't clear
what it does, but it *is modifiable*. It turns out it's connected to a
*hard-coded* Control/Cmd-Scroll action to scroll through the dims.

When scrolling, enough scroll needs to accumulate to trigger a slider
movement. Then it's reset back to 0 and repeats. The Control/Cmd key
needs to be held while scrolling, so what the `Reset scroll` keybinding
is doing is reseting this accumulated amount back to zero, if someone
stopped scrolling midway through a slice and then starts scrolling
again.

There is an issue though: The user *can't change the actual modifier
associated with the scroll action*, so having only the reset be a
changeable keybinding is confusing. If you change it, it effectively
does nothing because you won't trigger it when trying to scroll the
dims, which still triggers off Control/Cmd.
Further, doing this demonstrates that there is no issue, even with the
scroll action having a "memory" of past fractional scrolling. After
every step a reset already happens, so at most a fraction of a step when
starting again is just not noticeable. If you were last scrolling
forward, the next time you go to scroll you are slightly closer to the
next slice--or further away from the previous one--sure, but in practice
you don't notice it.

So I would argue that this confusing binding can just be removed. Which
is what I do here.
This means I have to update the settings test that was using the
keybinding because it was the first in the list, so I also did that.

Edit: using the first keybinding turned out to be in a lot of tests! And
it wasn't always obvious. I added comments to make it easier to follow
for the next time.

---------

Co-authored-by: Tim Monko <timmonko@gmail.com>