Reverse canvas grid order (#8053)
Alternative to #8044
Depends on #8057 
More context
https://github.com/napari/napari/pull/8044#issuecomment-3001025258 and
https://github.com/napari/napari/pull/7870#issuecomment-2965031040
Test with this script:

```python
import napari
from skimage import data

viewer, layer = napari.imshow(
        data.lily(),
        name='lily',
        channel_axis=2,
        colormap=['red', 'green', 'blue', 'gray'],
        blending='translucent',
        opacity=1,
        )

viewer.grid.enabled = True
viewer.grid.stride = 2

napari.run()
```

This PR reverses the default order of the panels with a default stride,
so that later layers end up to the bottom-right of earlier layers.

To enable the earlier behaviour, set a stride of -1, which can be set as
default in the viewer preferences.


![screenshot_4nFePQ5C@2x](https://github.com/user-attachments/assets/528aebca-d623-4f9a-97f4-691329d2a2a7)

## Script output:

main:


![screenshot_ikFxA97W@2x](https://github.com/user-attachments/assets/6d961f78-17ac-4ef7-aa8c-5fd66bc91dca)

this PR:


![screenshot_dlFFazIh@2x](https://github.com/user-attachments/assets/4516932d-0b2b-4e61-b378-6ff221d8c9be)

this PR with negative grid stride:


![screenshot_yQwsLxv6@2x](https://github.com/user-attachments/assets/4f791a10-9f00-4309-b029-50b6842f728d)

Co-authored-by: TimMonko <47310455+TimMonko@users.noreply.github.com>