Move export ROI and export figure implementations into `QtViewer` (#7950)
# References and relevant issues


# Description

In #7803 I got a few segfaults of `test_export_rois`. This PR refactors
`export_rois` and `export_figure` features into `QtViewer` and changes
tests to use the `qt_viewer` fixture instead of `make_napari_viewer`.
This should reduce the fragility of tests and allow the use of the
`export_rois` and `export_figure` features with qt_viewer embedded in a
custom window (as a nice additional outcome).

This PR also moves resizing canvas into a context manager to reduce code
duplication and provide better stability.

## Note

The error to easier count dots when reproducing 
```
   src\napari\_qt\_tests\test_qt_viewer.py ..........................s [3952/4910]
  ..Windows fatal exception: access violation
```

---------

Co-authored-by: Tim Monko <timmonko@gmail.com>
Co-authored-by: Wouter-Michiel Vierdag <w-mv@hotmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>