Allow use functions from PartSegCore-compiled-backend as numba alternative for data to texture mapping  (#6617)
# References and relevant issues

A way to fix warm up time reported here:
https://github.com/napari/napari/issues/3852#issuecomment-1905248322

# Description

This PR adds `PartSegCore-compiled-backend` as an alternative backend to
numba for fast mapping from raw data to texture.

Instead of depending on jited function, that have compilation delay it
uses precompiled code from `PartSegCore-compiled-backend`

It uses these two functions:

1.
[map_array_parallel](https://partsegcore-compiled-backend.readthedocs.io/en/stable/napari_mapping.html#PartSegCore_compiled_backend.napari_mapping.map_array_parallel)
2.
[zero_preserving_modulo_parallel](https://partsegcore-compiled-backend.readthedocs.io/en/stable/napari_mapping.html#PartSegCore_compiled_backend.napari_mapping.zero_preserving_modulo_parallel)

As it is implemented for only given data types, there is fallback to
numba also implemented (with warning).

Here is a notebook with performance measurement of these functions: 

https://github.com/4DNucleome/PartSegCore-compiled-backend/blob/master/notebooks/performance_napari.ipynb

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Tim Monko <timmonko@gmail.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>