Add 'zoom-box' to the viewer (#8004)
# References and relevant issues

This PR adds an interactive zoom box to the viewer canvas. 
This closes the original issue #7630 

# Description

It's only active in 2-D view and all you need to do is hold Shift,
press-down in the canvas, drag the mouse and release. Once released, a
`zoom` event is emitted by the `ViewerModel` which is processed in the
`canvas` class.

What's currently handled - things that need addressing by me:
- [x] add support for changing axes order
- [x] add controls for how the box is displayed - currently hard-coded
in the vispy code
- [x] improve how the zoom level is calculated - unlike in eg
matplotlib, we need to calculate the center of zoom and zoom scale. I
found this to be slightly harder to estimate so I've got code that kinda
works but could definitely do with improvement.
- [x] add tests  


https://github.com/user-attachments/assets/789407d6-aa4d-48d1-9938-df95e59ed118

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: Tim Monko <timmonko@gmail.com>
Co-authored-by: TimMonko <47310455+TimMonko@users.noreply.github.com>