Add automatic area and perimeter measurement for shapes + action (#7262)
# References and relevant issues
- Closes #4200
- Closes #3068

# Description
This is sort of an alternative to #6207, using the shapes layer instead
of an overlay. @pierrethibault might like this version better :)

Add a layer action that enables "measuring mode" on an existing shapes
layer (alternatively, we could spawn a new one).
It works by attaching a callback to `set_data` that automatically
calculates the perimeter and area of each shapes, sets them to columns
in the `features`, and shows that in the `text`.

The actual code is quite un-optimized, but as a proof of concept it
should do.



https://github.com/user-attachments/assets/e9aaa35c-6e08-4be9-8842-c1feda0c5825