Slightly increase pip test workflow timeout (#8143)
# References and relevant issues


# Description

I have observed multiple timeouts of our test suite on the pip job. The
test often takes around 25 minutes, so it is easy to hit the 30-minute
limit.

We should check if we could speed up the test suite itself, but let's
slightly increase the time limit to stop having unrelated crashes to the
PR content.