Add alpha to split rgba into channels and remerge image. (#8112)
# References and relevant issues
Closes https://github.com/napari/napari/issues/7806

# Description
This update maintains the alpha channel data when splitting and merging
RGBA images. Issue #7806 discusses
https://github.com/napari/napari/issues/7781, and I did not look back at
how these two issues are related.

@DragaDoncila I looked into this on the plane to pass the time and think
this PR implements what we talked through at the end of the day
yesterday. Also, I used the PyCharm debugger when troubleshooting and
updating the code and it was very helpful, so thanks for showing me how
to use that tool.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Tim Monko <timmonko@gmail.com>