Fix scalebar behaviour at high/low zoom (#8144)
# References and relevant issues
- closes https://github.com/napari/napari/issues/7956
- closes (kinda?) https://github.com/napari/napari/issues/6055

# Description

1. This issue has been bothering me for a long time and I finally
figured it out. On main, if you have the scalebar enabled and zoom out
beyond 1000, it will suddenly collapse to a single pixel.

Turns out, when the scalebar is dimensionless we need to ensure to keep
the world pixel length to the *real* lenght, because we don't have unit
multiples. Funny thing is, we already had this check in place, but for
some reason only applied to high zooms, and not low zooms. Anyways,
AFAICT this works.

2. The second issue is ultimately caused by our camera reporting a zoom
of `1` when it reaches a low enough scale factor. This code points to
https://github.com/napari/napari/issues/2875, but I can't figure out
why.

EDIT: it seems to be https://github.com/napari/napari/pull/2958. I tried
to remove this guard completely and I don't actually see that issue
anymore... I think it might be solved, either by the fact that we moved
a lot of camera updates to `on_draw` in
https://github.com/napari/napari/pull/7870 specifically because the
viewbox was degenerate, or for some other reason. Maybe we can just
remove this?

---------

Co-authored-by: Tim Monko <timmonko@gmail.com>