Use `pytest-rerunfailures` for flaky `test_toggle_fullscreen_from_maximized` test (#8151)
# References and relevant issues

# Description

Time to time yhe CI fails with `test_toggle_fullscreen_from_maximized`
on macOS.

It looks not related to our code but from a Windows manager on macOS
machines.

When looking for how to solve this, I found the `pytest-rerunfailures`
plugin that allows me to mark selected tests as flaky.

Maybe try this, but try not to extend the list of flaky tests too much.