Ensure contrast limits are computed on original dtype with projected thick slices (#8149)
# References and relevant issues
- closes #8148
- closes #8150 as well, apparently :P

# Description
Due to projections like `mean`, the dtype was changing during slicing.
Our thumbail code is a but more fragile, and so the contrast limits were
getting messed up.
The same is true for the contrast limits calculation for the data
itself, so a simillar fix is necessary.
The main issue happens with `np.uint8` data where we assume that clims
are `(0, 255)`, differently from any other dtype where clims are
calculated based on the available data itself, or set to `(0, 1)`.

The fix is to use the original, unprojected dtype of the layer data to
calculate clims, instead of the (potentially) new dtype from the
projected slice.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>