feat: Add option to execute python code by drag'n'drop script to viewer (#8135)
# References and relevant issues

# Description

When reviewing possible materials for tutorial days on EuroSciPy. I spot
that we need to either prepare a fully functional plugin or control
napari from a Jupyter Notebook (or other interactive session).

We may also start napari from script. 

But why not drag and drop a script into the viewer? 

This will allow to use bundle, but also fast prototype of simple
plugins.

What did you think?


## Disclaimer

It is done using a reader plugin for now, but I think that it would be
better to put such logic earlier in the chain. But, if you are happy
with the idea I prefer it merge and add to 0.6.3 to have it in stable
for tutorials.

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: Carol Willing <carolcode@willingconsulting.com>