Update Citation file for 0.6.1 through 0.6.3 (#8138)
The release manager (me) has dropped the ball adding new contributors to
the citation file for 0.6.2 and 0.6.3. This PR adds all new contributors
for the last two releases. I have also moved Carol to the core team
section.

@jacopoabramo @Llewi @rahul713rk @ahuang11 @FilBalza @ian-coccimiglio
@8bitbiscuit @rwkozar,
if you would like to appear in our [Zenodo
record](https://zenodo.org/records/16351526) (a citable doi reference
https://doi.org/10.5281/zenodo.3555620), please reply here with the
following info in a code block like so:


```
- given-names:
  family-names:
  affiliation:
  orcid: (optional)
  alias: (your GH username)
```

and I'll add it to the file. Thank you!