Prevent `napari.run` from being executed when running scripts from napari (#8187)
# References and relevant issues

Will solve
https://github.com/napari/napari/pull/8185#issuecomment-3141371791

# Description

Here I add monkeypatching of `_ipython_has_eventloop` to early stop
execution of `napari.run()`
This prevents the creation of nested even loops. It is required to solve
the problem pointed out in the linked comment.

Also, this allows us to set `__name__` in the executed script to
`'__main__'` which enables execution of code under `if __name__ ==
'__main__':`, allowing us to have working `multiple_viewer_widget.py`
after drag and drop.

---------

Co-authored-by: Carol Willing <carolcode@willingconsulting.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>