Fix slider label shifted down, by overrwite QLineEdit qss rules (#8184)
# References and relevant issues

https://github.com/napari/napari/pull/7355#issuecomment-3134232094

# Description

After long debugging, I found that the slider label is shifted because
of 2px padding of QLineEdit set in our QSS. It was a long and not
obvious debug route.


So it means that the problem cannot be solved by a patch of superqt and
will be with us until the next napari release.

---------

Co-authored-by: Tim Monko <timmonko@gmail.com>