Use global register for units in ScaleBar (#8226)
# References and relevant issues


https://github.com/brisvag/napari-tutorial-euroscipy2025/blob/460531e31d98ee7669b2bd501a9ba20ad56e906c/presentation.py#L4

# Description

During preparation of the EuroSciPy tutorial, we spot that ScaleBar is
using its own register of units instead of a global one.

This is different from my job for units in layers, hidden under private
API. It also means that all packages that want to use units compatible
with napari then need to import from napari instead of just using pint.