Temporary increase timeout of macOS jobs (#8237)
# References and relevant issues

# Description

In recent times we often see tests timing out in CI. We collect JSON
reports with timing information that are expected to help solve this
problem, but unfortunately it looks to be to big to load to existing
viewer (see
https://github.com/Carreau/pytest-json-report-viewer/issues/8)

My next plan is to open a separate PR with running only a subset of
tests to try to identify where the problem is. But for this time we
should increase the timeout because double execution of a 40-minute run
is longer than just one 50 minutes.