Update Qt backends used in tests to test more on Qt6 than Qt5 (#8229)
# References and relevant issues

# Description

In recent times we started shifting to make Qt6 binding our default one.
We also plan to drop PySide2 shortly.

So it is time to depend more on Qt6 in our test matrix. 

Currently, PySide6 is a fragile one and bundled with PyQt5. It means
that a failure in PySide6 requires rerunning the whole PyQt5 test step.
This PR gives PySide6 its own job. I do not see which test may be
removed from our cases. So it is one additional job. Will try to fix.

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>