[maint] Revert back to windows-2022 for windows runners (#8265)
# References and relevant issues
Windows jobs have all been failing, which seems to be connected to
`windows-latest` runners being updated:

https://github.blog/changelog/2025-07-31-github-actions-new-apis-and-windows-latest-migration-notice/#windows-latest-image-label-migration

# Description
Changes from using `windows-latest` which is now windows server 2025, to
using `windows-2022` which is what `-latest` used to use.

This should only be a temporary change, to get CI to work again while we
work on a workaround that works on `-latest` in a branch.