Fix effect of scaling when converting shapes to labels (#8098)
# References and relevant issues
Closes https://github.com/napari/napari/issues/7962.

# Description
The x and y dimensions of the label went from zero to the maximum x and
y dimensions of the shape, rather than the image. This PR sets the
maximum x and y label dimensions of the label equal to world maximum x
and y values.

New issues for consideration: 

The `to_layers` function in `src/napari/layers/shapes/shapes.py` assumes
that the minimum world x and y values are zero. This assumption is
initially valid, but the user can make adjustments that cause the
minimum x and/or y values to become non-zero. When this occurs, regions
of the canvas with a negative x or y value are not accessible in the
label. Fully resolving the intent of issue 7962 would require applying
both the minimum and maximum coordinates of the canvas to the label.

The label extent is static after creation. If a label is created from a
shape when the world dimensions are 400x400, if a future action cause
the canvas size to increase (e.g. to 800x800), the label shape remains
400x400 and regions of the image outside the initial label extent are
inaccessible.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Tim Monko <timmonko@gmail.com>
Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: Peter Sobolewski <pete.sd@gmail.com>