Handle affine layer metadata when splitting RGB images (#8256)
Closes: https://github.com/napari/napari/issues/7085

# Description
As noted in the issue, `split_rgb` was using `stack_to_images`, which
throws out the affine information (and rotate and shear).
In fact, handling those for RGB isn't too bad because RGB images have a
fixed *non-spatial axis* and are already handled as 2D by napari. As
noted by Andreas, most of the logic was already implemented in
`split_channels`, but that function wasn't being used for split_rgb.

So, this PR switches `split_rgb` to use `split_channels` which requires
a minor adjustment for the differences in arguments and returns, but
otherwise is straightforward. I also ensure that shear and rotate are
handled.
For the test, I implement the test from Andreas in the issue
(Co-authored).
I also tested locally with transform mode and everything works quite
nicely!

Note: this issue is more pressing now that image layers have the
Transform mode which uses Affine.

---------

Co-authored-by: Andreas Eisenbarth <andreas.eisenbarth@embl.de>
Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>