use `Selection` class in shapes layer (#8297)
Closes #6886

# Description
This PR introduces the use of the pysignal `Selection` class to further
standardize the events across different layers. In particular, this
brings the use of `layer.selected_data.events...` to the shapes layer,
as request by some users.

- [x] My PR is the minimum possible work for the desired functionality
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to docstrings and documentation
(open a PR on the docs repository (https://github.com/napari/docs) if
relevant!)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>