Use single settings path for all `uv tool run` (#8250)
# References and relevant issues


https://napari.zulipchat.com/#narrow/channel/212875-general/topic/.E2.9C.94.20global.20settings/near/537877859

https://docs.astral.sh/uv/concepts/tools/#tool-environments

# Description

This PR adds a check if the current environment is created by `uvx` (or
`uv tool run`), if yes, then disable calculation of `sha` of path so
recreation of venv does not cause selection of different settings path.

<img width="624" height="220" alt="image"
src="https://github.com/user-attachments/assets/19fd5395-852f-4549-b71d-46133c91040a"
/>

To avoid settings conflicts, we use the napari version in path. If there
are no settings, but settings exist for older napari versions, we use
the most modern as the default.

This does not solve #8246

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>