Fix invisible zoom box (#8344)
# References and relevant issues
- closes https://github.com/napari/napari/issues/8342

# Description
This slipped through when refactoring in #7836. The root of the problem
is:
- canvas overlays can have "locked" position (top_left, bottom_right,
etc), or "free" positions (whatever the overlay wants)
- overlays need to opt out of the "locked" mode, which takes care of
tiling and auto-sets the position
- to opt out you need to specifically set `overlaymodel.position` to a
value that *isn't* a `CanvasPosition` enum
- you also should override the method `_on_position_change` in the
`VispyOverlay` class in order to avoid unnecessarily updating the tiling

I changed the naming scheme to adapt to this, and now everything works.
Basically what was happening is that the overlay was being shifted to
`bottom_right` automatically, which meant it was outside of the canvas
:P