Fix shape selection in a single plane when shapes are on multiple planes (#8335)
# References and relevant issues
Closes https://github.com/napari/napari/issues/8293

# Description
In this PR I implement 3 main things to address the issue:
1. a guard to ensure _selected_box is not None before trying to index
into it
2. use _data_view.displayed_vertices_to_shape_num as a mask on
displayed_vertexes to get the vertexes of shapes in the view slice. The
issue with the current `self._data_view._vertices_slice_available(i)` is
that it returns vertex indexes from all vertexes rather than just the
subset of the displayed view slice.
3. rename the keybinding function from `select_all_shapes` to
`select_shapes_in_slice`, which is what it does. Then I ensure that
behavior, replacing the selection rather than appending. See related
Points issue: https://github.com/napari/napari/issues/8333 This prevents
trying to draw an interaction_box across slices.

Finally, I add a test (with help from gemini) for selecting within a
slice of 3d shapes (also change from using random data, just in case).
This test fails on main.

Tests pass for me locally and selection of Shapes appears to work
correctly when I play with it in napari, but it would be good to ensure
someone pulls this down and checks in case I missed something!

Performance note: the issue was introduced in a performance-related PR
(https://github.com/napari/napari/commit/2f666ca27f5e2735a826ebadf8a0788211c0037a).
I tested this branch and main with a 150K shape layer and selection
seems on par if not slightly faster with this PR.