Fix vispy error traceback (#8346)
# References and relevant issues
- closes https://github.com/napari/napari/issues/4570 (and likely
similar issues I missed)

# Description
When something goes wrong in vispy, the resulting traceback is usually
very long and messy and full of info that's useless for users, to the
point that when users ask for help, the relevant information from the
traceback is often buried so far that they don't even copy it to the
issue.

This PR attempts to mitigate this in 2 ways:
- ignoring "useless" errors, which have no useful info in the traceback
and are always caused by another previous unresolved error, but end up
burying it
- adding some extra info for known issues that are hard to decipher.

I'm tempted to add the `cannot size object` issue to the ignore list,
but I'm not 100% sure that this won't hide actual useful information. I
think it's better to be conservative here and if it shows up again we
can update this ignore list as needed.