Remove deprecated `napari.view_*` methods (#8337)
# References and relevant issues

Closes #8082 and #7047
Depends on napari/docs#864

# Description

This PR is spun out, in a sense, from #8334 because instead of updating
so much code, we can just deprecate the functions.
It also serves to show the value of actually removing the view methods,
and help make the decision for if we want to fully deprecate.

This deprecates all the `napari.view_*` methods, and attempted to check
the repo for all instances of usage and clean it up.
Note how by keeping #8334 we have to do quirky things with docstring
parser again. So, by deprecating we avoid having to modify these
methods, instead removing.

I think `view_layers.py` may be renamed to something else like
`imshow.py` but I didn't want to be adjusting any python paths, for now.