Add a seed argument to built-in samples with random seeds (#8317)
# References and relevant issues

Follow-up to #8097. Closes #7131 

# Description

Adds seed as a argument to functions which create the `Balls` ,
`Balls3D`, `Blobs`, and `Blobs3D` samples. Also adds default shape arg
to balls samples. This is an improvement over previous implementation
where the seed and shape was hard set. Now we have the advantage of the
sample being consistent on normal call of the function, but users can
also set a different seed if desired.

This came about because I was hasty in my inital review during SciPy
sprints. Keeps popping in my head when I'm away from my desk... so I'm
finally making this PR.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>