initialize label selection spinbox to a correct value (#8382)
# References and relevant issues
closes #8380 

# Description
Before #8274 we called the callback to initialize the label spinbox to
the current label value, but after #8274 the spinbox was left to a
default state. This PR updates the current label value before the
callback is called.
DragaDoncila and I checked the rest of the widgets, but we found no
issues, as the current value is either initialized or 0 by default.
Added a test that fails on main but passes on this PR.