Fix overlay initialization and layer addition slowdown (#8443)
# References and relevant issues
- follow up #8423
- fix #8434
- fix #8421

# Description
When trying to improve #8421, I caused #8434. This prompted me to
finally rework the overlay initialization machinery to be less wasteful.
This PR has 3 main changes compared to main:

1. Instead of deleting and recreating all overlays every time an
scenegraph update is necessary, vispy overlays are reused and simply
reparented (also their duplicates when working in gridded mode). I could
not make this work in the past for some reason, but now it seems to work
just fine.
2. VispyLayerOverlay initialization is deferred until the overlay
becomes visible the first time. This removes the hidden cost of layer
overlays which was causing extra slowdown when creating new layers.
3. Only regenerate the grid when necessary (when the grid size changes),
which drastically speeds up layer addition for all cases where the grid
does not change shape.

This results in the following profiles compared to #8421:

<img width="1374" height="653" alt="newplot2"
src="https://github.com/user-attachments/assets/afbd2448-5e37-4693-8aec-ab4d12db5a31"
/>

Note how now in grid mode the baseline is much better, and there's those
spikes corresponding to when the grid needs to be made bigger.